+++ tests/lib_linker_flags/lib_mod1/lib_mod1/api/mod1.h
@@ -1,0 +1,4 @@
+#ifndef MOD1_H
+#define MOD1_H
+void mod1();
+#endif