+++ tests/lib_linker_flags/lib_mod1/lib_mod1/lib_build_info.cmake
@@ -1,0 +1,7 @@
+set(LIB_NAME lib_mod1)
+set(LIB_VERSION 1.0.0)
+set(LIB_INCLUDES api)
+set(LIB_DEPENDENT_MODULES "lib_mod1_1")
+set(LIB_LINKER_FLAGS -v)
+
+XMOS_REGISTER_MODULE()