+++ tests/lib_linker_flags/lib_mod1_1/lib_mod1_1/api/mod1_1.h
@@ -1,0 +1,4 @@
+#ifndef MOD1_1_H
+#define MOD1_1_H
+void mod1_1();
+#endif